feature/01: Associando imovel ao endereco, utilizando feigh e configurações do mongo

+++ api-imobiliaria/src/main/java/com/br/imobiliaria/api/imovel/repository/dto/ImovelResponseDto.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ImovelResponseDto {
-    private Integer idImovel;
+    private Long idImovel;
     private String tipoImovel;
     private String subTipoImovel;
     private String cidade;
